chore: pre-release cleanup — mypy clean (404→0), bug fixes, R validation harness, docs
- mypy: 404 → 0 errors in 138 source files (no behavior changes except listed fixes)
- Fix ZINB dead NB init (family_params kwarg silently swallowed)
- Fix plot() diagnostics TypeError for GLM results
- Fix plot_smooth_effect silent polynomial fallback (design_matrix → basis_matrix, correct knot count)
- Fix anova_glm TypeError on models without deviance (fallback to LRT)
- Declare GLMResult.theta_, CovarianceKind alias, GAMMResult.random_effects nested dict
- R validation: sys.executable, digits=15, Gamma AIC under R's φ=D/n convention → 20/20 pass
- Docs: theta='estimate' section in glm.rst, VALIDATION.md Gamma dispersion note,
  fix stale family_params= examples in family docstrings

+++ aurora/smoothing/splines/cubic.py
@@ -201,8 +201,9 @@ class CubicSplineBasis:
 
         def _D(p: int, q: int) -> float:
             """Integral of d_p''(x) d_q''(x) over the real line."""
-            return self._integrate_second_derivatives(all_knots[p], all_knots[q], upper) / (
-                (upper - all_knots[p]) * (upper - all_knots[q])
+            return float(
+                self._integrate_second_derivatives(all_knots[p], all_knots[q], upper)
+                / ((upper - all_knots[p]) * (upper - all_knots[q]))
             )
 
         # For cubic terms: S[i+2, j+2] = D(i,j) - D(i,K-1) - D(K-1,j) + D(K-1,K-1)
